refactor(middleware): move parseFormData from middleware to utils/helper
Consolidate parseFormData middleware into utils/helper.js for better code organization and reusability
